Moved files into new folder pages/home/
