+++ src/App.js
@@ -4,21 +4,21 @@ import { BookstorePage } from './components/pages/bookstore/BookstorePage'
 import { Register } from './components/auth/Register';
 import { ForgotPassword } from './components/auth/ForgotPassword';
 import { Login } from './components/auth/Login';
-import { BestSellers } from './components/bestSellers/BestSellers'
-import { BestSellersData } from './components/bestSellers/BestSellersData';
+import { Home } from './components/pages/home/home';
 
 import './App.css';
 
 export const App = () => {
   return (
     <div className="App">
-      <Contact />
-      <BestSellers />
-      <BestSellersData />
-      <BookstorePage />
-      <Register />
-      <ForgotPassword />
-      <Login />
+      {/* <Contact /> */}
+      {/* <BestSellers /> */}
+      <Home />
+
+      {/* <BookstorePage /> */}
+      {/* <Register /> */}
+      {/* <ForgotPassword /> */}
+      {/* <Login /> */}
       <Footer />
     </div>
   );
