+++ src/components/pages/home/home.css
@@ -1,0 +1,1 @@
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400..700&display=swap');
