+++ src/components/pages/home/home.js
@@ -1,0 +1,11 @@
+import {LoginOrRegister} from './LoginOrRegister'
+import {BestSellers} from './BestSellers'
+
+export const Home = () => {
+    return (
+        <>
+            <LoginOrRegister />
+            <BestSellers />
+        </>
+    )
+}
